# docs/swagger.yaml
definitions:
  model.BasicResponse:
    properties:
      message:
        type: string
      result: {}
      status:
        type: boolean
    type: object
...
  model.NFTInfo:
    properties:
      nft:
        type: string
      nft_file_name:
        type: string
      nft_metadata:
        type: string
      nft_value:
        type: number
      owner_did:
        type: string
    type: object
  model.NFTList:
    properties:
      message:
        type: string
      nfts:
        items:
          $ref: '#/definitions/model.NFTInfo'
        type: array
      result: {}
      status:
        type: boolean
    type: object
...
  server.DeployNFTSwaggoInput:
    properties:
      did:
        type: string
      nft:
        type: string
      nft_data:
        type: string
      nft_file_name:
        type: string
      nft_metadata:
        type: string
      nft_value:
        type: number
      quorum_type:
        type: integer
    type: object
...
  server.ExecuteNFTSwaggoInput:
    properties:
      comment:
        type: string
      executor:
        type: string
      nft:
        type: string
      nft_data:
        type: string
      nft_value:
        type: number
      quorum_type:
        type: integer
      receiver:
        type: string
    type: object
...
  server.NewNFTSwaggoInput:
    properties:
      nft:
        type: string
    type: object
...
info:
  contact:
    email: murali.c@ensurity.com
    name: API Support
  description: Rubix core API to control & manage the node.
  title: Rubix Core
  version: "0.9"
paths:
...
  /api/create-nft:
    post:
      consumes:
      - multipart/form-data
      description: This API will create new NFT
      parameters:
      - description: DID
        in: formData
        name: did
        required: true
        type: string
      - description: JSON file which contains information about the NFT
        in: formData
        name: metadata
        required: true
        type: file
      - description: File which is meant to be an NFT
        in: formData
        name: artifact
        required: true
        type: file
      produces:
      - multipart/form-data
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BasicResponse'
      summary: Create NFT
      tags:
      - NFT
  /api/deploy-nft:
    post:
      consumes:
      - application/json
      description: This API will deploy the NFT
      operationId: deploy-nft
      parameters:
      - description: Deploy nft
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/server.DeployNFTSwaggoInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BasicResponse'
      summary: Deploy NFT
      tags:
      - NFT
...
  /api/execute-nft:
    post:
      consumes:
      - application/json
      description: This API will add a new block which indicates either transfer of
        ownership of NFT or internal state change through self-execution
      parameters:
      - description: 'Transfer the ownership of particular NFT or self-execution with
          some data if ''receiver'' is empty '
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/server.ExecuteNFTSwaggoInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BasicResponse'
      summary: Execution of NFT
      tags:
      - NFT
...
  /api/fetch-nft:
    get:
      consumes:
      - application/json
      description: This API will Fetch NFT
      operationId: fetch-nft
      parameters:
      - in: query
        name: nft
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BasicResponse'
      summary: Fetch NFT
      tags:
      - NFT
...
  /api/get-nfts-by-did:
    get:
      consumes:
      - application/json
      description: This API will get all NFTs owned by the particular did
      parameters:
      - in: query
        name: did
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.NFTList'
      summary: Get NFTs owned by the particular did
      tags:
      - NFT
...
  /api/subscribe-nft:
    post:
      consumes:
      - application/json
      description: This API endpoint allows subscribing to a NFT.
      parameters:
      - description: Subscribe to input nft
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/server.NewNFTSwaggoInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BasicResponse'
      summary: Subscribe to NFT
      tags:
      - NFT
...
securityDefinitions:
  SessionToken:
    in: header
    name: Session-Token
    type: apiKey
swagger: "2.0"